refactor: RankTierView -> RankTierCell로 변경
