+++ mejai/mejai/Presentation/Scenes/Home/Views/RankTierView.swift
@@ -7,34 +7,33 @@
 
 import UIKit
 
-enum RankTierType: String {
-    case flex = "자유 랭크"
-    case solo = "솔로 랭크"
-}
-
 final class RankTierView: UIView {
     // MARK: - Components
     
     private let titleLabel = {
         let label = UILabel()
-        label.text = "자유 랭크"
-        label.applyTypography(with: .caption2)
-        label.textColor = .gray04
-        return label
-    }()
-    
-    private let tierImageView = {
-        let imageView = UIImageView()
-        imageView.backgroundColor = .gray01
-        return imageView
-    }()
-    
-    private let tierLabel = {
-        let label = UILabel()
-        label.text = "PLATINUM II"
-        label.applyTypography(with: .body3)
+        label.text = "티어"
+        label.applyTypography(with: .title3)
         label.textColor = .gray09
         return label
+    }()
+    
+    private lazy var collectionView = {
+        let layout = UICollectionViewFlowLayout()
+        layout.minimumInteritemSpacing = 15
+        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
+        collectionView.register(cellType: RankTierCell.self)
+        collectionView.backgroundColor = .backgroundSecondary
+        collectionView.layer.cornerRadius = Constants.Radius.lg
+        collectionView.delegate = self
+        collectionView.dataSource = self
+        return collectionView
+    }()
+    
+    private let line = {
+        let view = UIView()
+        view.backgroundColor = .gray01
+        return view
     }()
     
     // MARK: - Init
@@ -54,21 +53,41 @@ final class RankTierView: UIView {
     private func configureLayout() {
         addSubview(titleLabel)
         titleLabel.snp.makeConstraints { make in
-            make.top.leading.trailing.equalToSuperview()
+            make.top.leading.equalToSuperview()
         }
         
-        addSubview(tierImageView)
-        tierImageView.snp.makeConstraints { make in
-            make.width.height.equalTo(Constants.Layout.IconImage.small)
+        addSubview(collectionView)
+        collectionView.snp.makeConstraints { make in
             make.top.equalTo(titleLabel.snp.bottom).offset(Constants.Spacing.xs)
-            make.leading.equalToSuperview()
+            make.leading.bottom.trailing.equalToSuperview()
+            make.height.equalTo(75)
         }
         
-        addSubview(tierLabel)
-        tierLabel.snp.makeConstraints { make in
-            make.top.equalTo(tierImageView)
-            make.leading.equalTo(tierImageView.snp.trailing).offset(Constants.Spacing.xs)
-            make.trailing.equalToSuperview()
+        addSubview(line)
+        line.snp.makeConstraints { make in
+            make.width.equalTo(1)
+            make.height.equalTo(collectionView).inset(Constants.Spacing.sm)
+            make.centerX.equalToSuperview()
+            make.centerY.equalTo(collectionView)
         }
+    }
+}
+
+extension RankTierView: UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        let width = (collectionView.frame.width - 15) / 2
+        return CGSize(width: width, height: collectionView.frame.height)
+    }
+}
+
+// TODO: diffable data source로 바꿀 것
+
+extension RankTierView: UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        2
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        return collectionView.dequeueReusableCell(for: indexPath, cellType: RankTierCell.self)
     }
 }
